core/pyproject.toml
[project]
name = "minima"
version = "0.1.0"
description = "Minima LMS"
authors = [{ name = "Cobel", email = "cobel1024@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14,<3.15.0"
dependencies = [
  "django",
  "django-storages[s3]",
  "django-pgtrigger",
  "django-taggit",
  "django-treebeard",
  "django-redis",
  "django-phonenumber-field[phonenumberslite]",
  "django-unfold",
  "django-opensearch-dsl",
  "psycopg[binary]",
  "passlib[bcrypt]",
  "tika",
  "pillow",
  "django-import-export",
  "celery",
  "django-celery-results",
  "uvicorn",
  "pyjwt",
  "pyotp",
  "qrcode",
  "django-cleanup",
  "django-pghistory",
  "django-ipware",
  "ua-parser[regex]",
  "django-otp",
  "django-ninja",
  "msgspec",
  "pydantic[email]",
  "isodate",
  "django-jsonform",
  "beautifulsoup4",
  "google-genai",
  "fpdf2",
  "pypdfium2",
  "webvtt-py",
  "authlib",
  "httpx",
  "firebase-admin>=7.1.0",
]

[dependency-groups]
dev = [
  "typer",
  "pytest-django",
  "pytest-xdist",
  "pytest-cov",
  "pytest",
  "factory-boy",
  "mimesis",
  "pytest-order",
  "pytest-mock",
  "boto3-stubs[s3]",
  "yt-dlp",
  "youtube-transcript-api",
  "mjml-python",
  "openpyxl",
  "django-stubs",
  "typing-extensions",
]

[tool.django-stubs]
django_settings_module = "minima.settings"

[tool.ruff]
line-length = 120
extend-exclude = ["**/migrations/**", "typings"]

[tool.ruff.format]
preview = true
skip-magic-trailing-comma = true

[tool.djlint]
profile = "django"
blank_line_after_tag = "load,extends,include"

[tool.ty.src]
exclude = ["**/migrations/**", "**/tests/**"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "minima.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
markers = ["e2e: End-to-end tests"]
addopts = "--tb=short --strict-markers -m 'not e2e'"
cache_dir = "/tmp/.pytest_cache"

[tool.coverage.run]
data_file = "/tmp/.coverage"

[tool.coverage.html]
directory = "/tmp/htmlcov"

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = [
  "**/node_modules",
  "**/__pycache__",
  "**/.venv/**/*",
  "**/migrations/**/*",
]

[tool.pyrefly.errors]
inconsistent-inheritance = false
bad-override = false
